orders-svc/requirements.txt
Flask==2.3.3
kafka-python==2.0.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0